Rip out the inflector's cache
The cache contains a race condition between the time when a process
checks for a registered process called "re_cache" and the time that
it creates and registers the process itself. If a cache is to exist
at all, it needs to follow OTP principles. In the meantime I am ripping
it out.