Fix issues with riak_search
The main issue was that the proplist (that is written to Riak) needs to have keys and values encoded as binary, in order for riak_search to index stuff.

This means that [{foo, "bar"}] will not be index, whereas [{<<"foo">>, <<"bar">>}] will be.

Since this change broke the "internal" (boss_db) search, this was fixed, too.

TODO: After the call to riakc_pb_socket:search we already have the entire object. find_acc fetches it again based on the ID.