Introduce new before_update/1 and after_update/1 hooks
The new before_update/1 and after_update/1 hooks behave the same way
as before_update/0 and after_update/0. The difference is that they
do provide access to the old boss record as an additional parameter.

This addition is fully backward compatible due to the fact that the
before_update/0 and after_update/0 hooks act as fallbacks.
For example, if both hooks before_update/0 and before_update/1 exists
in a boss model module than before_update/1 will be prioritized and
before_update/0 will be ignored. Same behaviour applies
to after_update/1 and after_update/0.